Move scripts to src/ and update Dockerfile
Move scripts as in lab 1, so it has similar structure
